StarterTemplate/requirements.txt
asgiref==3.10.0
Django==4.2.25
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.16.1
dnspython==2.7.0
dotenv==0.9.9
mongoengine==0.29.1
pymongo==3.11.4
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
pycryptodome==3.23.0
pyotp==2.9.0
channels==4.3.1
daphne==4.2.1
